Midi_Scanner: s_diffValues int -> size_t

+++ FretFrameworkV2/Song Library/Entry/Midi_Scan_Drums.h
@@ -71,7 +71,7 @@ void Midi_Scanner_Instrument::Scanner<DrumNote_Legacy>::parseLaneColor(MidiNote 
 	if (lane == 5)
 		m_scan.setType(DrumType_Enum::FIVELANE);
 
-	const int diff = s_diffValues[noteValue];
+	const size_t diff = s_diffValues[noteValue];
 	if (!m_difficulties[diff].active)
 	{
 		if constexpr (!NoteOn)
